FEAT annotate html capture and automatic click func
