+++ public/manifest.json
@@ -16,7 +16,8 @@
     {
       "matches": ["<all_urls>"],
       "js": ["content.js"],
-      "run_at": "document_end"
+      "run_at": "document_end",
+      "type": "module"
 
     }
   ],
